extract modal logic into file
